% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jamfuncs.R
\name{fun3}
\alias{fun3}
\title{Test function level 3}
\usage{
fun3(var31, var32, var33, ...)
}
\arguments{
\item{var31}{first variable for function 3}

\item{var32}{second variable for function 3}

\item{var33}{third variable for function 3}
}
\description{
Test function level 3
}
